Rewrite alias chain collapse logic
This make things a bit simpler, and also makes it more straightforwad to
test. Previously was written in such a way that it assumed you had
checked to make sure an alias existed before calling it. I've re-written
it (and its calling methods) to remove that assumption; it now interally
handles aliases that do not exist in a better way.