Make Brief ingestion behavior better, more clear
Although `make()` always accepted Briefs as arguments, this allows `new`
to do so as well, albeit with slightly different behavior. This behavior
is tested (currently 100% test coverage) and documented, but should be
kept in mind when using the library.

This also introduces the public `export()` method and the protected
`import()` method, which allow you to directly copy a Brief's innards.
They generally shouldn't be used when working with Briefs, but
`export()` has to be public in order for Briefs to be able to ingest one
another when using `new`.