Remove internal "arguments" terminology
The original version of much of the interal Brief documentation used
"argument(s)" to refer to data stored on the Brief. This was influenced
by some of the original intent, but I've come to feel that--especially
with the introduction of more callables--this terminology serves only to
confuse the documentation.

Hence I have switched to using, simply, "data" to describe the data
stored in a Brief, and "store" to refer to the location where that data
is stored.