Make callables nicer to work with
This implements a simple container for callables that requires us to do
way fewer checks and tests when working with them.

It also eliminates the rather weird internal practice of allowing for
`true` as a valid callable over the life of the objects: Now it checks
for that value on instantiation and then creates an anonymous function
to handle that, which simplifies those interactions a great deal.

This closes #9